pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.5",
    "django-import-export[xlsx]>=4.3.5",
    "django-taggit>=6.1.0",
    "markdown>=3.7",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "faker>=35.0.0",
    "ruff>=0.9.2",
]

[tool.ruff]
line-length = 100

[tool.djlint]
profile="django"
blank_line_after_tag="load,extends,include"
blank_line_before_tag="load,extends,include"
max_line_length=110
